// partner-search/frontend/src/app/api/model/branch-dto.ts
/**
 * traefik-microstack API Gateway
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Branch office information containing address and contact details
 */
export interface BranchDto {
  /**
   * Unique branch identifier - primary key
   */
  branchNumber: number;
  /**
   * Postal code of branch location
   */
  postalCode?: string;
  /**
   * Branch designation/name
   */
  designation?: string;
  /**
   * Street address
   */
  street?: string;
  /**
   * City name
   */
  city?: string;
  /**
   * Phone number
   */
  phone?: string;
  /**
   * Fax number
   */
  fax?: string;
  /**
   * Last modification date
   */
  lastModified?: string;
  /**
   * State/Federal state code
   */
  state?: string;
  /**
   * Region number
   */
  regionNumber?: number;
  /**
   * Post office box
   */
  poBox?: string;
  /**
   * Post office box postal code
   */
  poBoxPostalCode?: string;
  /**
   * Email address
   */
  email?: string;
}